+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - dumper: support the `{class}` variable in config's `path` param.
 - dumper: don't dump large messages, configurable by `max_dump_size` param (64KiB by default).
 - telemeter: add `Retention::ResetOnScrape` policy as the simplest way to protect against stabilization with time.
+- scope: `try_set_trace_id`.
 
 ### Changed
 - telemeter: use `Retention::ResetOnScrape` by default.